LLVM Fix Pass: Find correct insertion point
Before, missing Store Instructions would be inserted
just before the first load. This breaks in cases, where
the load is nested inside loops, because the value of the
variable would then be reset in each iteration.

Now, we find the first Alloca Instruction (declaration) for
the variable and insert the Store Instruction after that.
There may be some problems with multiple Alloca Instructions
for a single variable, but I don't we will ever have to deal with it.

This commit makes all the IR Fix tests pass!